+++ CHANGELOG.md
@@ -19,6 +19,16 @@ The versions follow [semantic versioning](https://semver.org).
 
 ## Unreleased
 
+### Changed
+
+- `Project.reuse_info_of` now extracts, combines and returns information both
+  from the file itself and from debian/copyright.
+
+- `ReuseInfo` now holds sets instead of lists.
+
+  - As a result of this, `ReuseInfo` will not hold duplicates of copyright lines
+    or SPDX expressions.
+
 ## 0.1.1 - 2017-12-14
 
 ### Changed